requirements.txt
# Django & REST framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<25.0

# Database
psycopg2-binary>=2.9,<3.0

# Async task queue
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-results>=2.5,<3.0

# HTTP client (Bolna API)
requests>=2.31,<3.0

# AWS S3 / MinIO
boto3>=1.34,<2.0

# Environment & config
python-dotenv>=1.0,<2.0

# Production server
gunicorn>=21.2,<23.0
